Add list support for all renderer modes, add caching
This introduces metadata information about all included pages and adds a
cache handler that compares the list of actually included pages with the
list of pages that would be included. Furthermore the modification time
is checked for all included pages if any page in the wiki was modified
more recently than the cache file.