Add caching for extracted content from a page
This adds a cache for the whole extracted content from a page, this
prevents reading and parsing the instructions multiple times for
metadata and xhtml and if the same page appears twice in a page.
Furthermore if only one of the included pages is changed the other
caches can still be used.